requirements.txt
accelerate==1.1.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==24.2.0
audioread==3.0.1
auto_gptq==0.7.1
autopep8==2.3.1
av==14.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
bertviz==1.4.0
black==24.10.0
bleach==6.2.0
boto3==1.37.21
botocore==1.37.21
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
datasets==3.1.0
decorator==5.2.1
deepspeed==0.16.4
defusedxml==0.7.1
dill==0.3.8
docopt==0.6.2
docstring_parser==0.16
einops==0.8.1
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.13.1
fire==0.7.0
flash_attn==2.7.4.post1
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.2.0
gekko==1.2.1
google==3.0.0
gradio==5.25.0
gradio_client==1.8.0
groovy==0.1.2
h11==0.14.0
hf_transfer==0.1.8
hjson==3.1.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
ipython==8.12.3
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.2.1
ninja==1.11.1.4
nltk==3.9.1
numba==0.61.2
numpy==1.26.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.3.2
openprompt==1.0.1
openpyxl==3.1.5
optimum==1.23.3
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
peft==0.15.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.50
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.0.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rouge==1.0.0
rouge-chinese==1.0.3
rpds-py==0.24.0
ruff==0.11.5
s3transfer==0.11.4
safehttpx==0.1.6
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.15.0
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.1.96
shellingham==1.5.4
shtab==1.7.2
six==1.16.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.46.2
swankit==0.1.6
swanlab==0.5.3
sympy==1.13.1
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
transformers==4.49.0
triton==3.1.0
trl==0.9.6
typeguard==4.4.2
typer==0.15.2
typing_extensions==4.13.0
tyro==0.8.14
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.1
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
yarg==0.1.9
yarl==1.17.2
